pyproject.toml
[project]
name = "coursework3"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.16.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
